% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlFunctions.R
\name{add_new_property}
\alias{add_new_property}
\title{Add a new property to xml object}
\usage{
add_new_property(xml, name, value, type)
}
\arguments{
\item{xml}{object}

\item{name}{new property name}

\item{value}{value to assign}

\item{type}{type of property in c("lpwstr","bool","i4","filetime")}
}
\value{
Nothing
}
\description{
Add a new property to xml object
}
