% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlFunctions.R
\name{get_names}
\alias{get_names}
\title{Get the names of the document properties}
\usage{
get_names(xml)
}
\arguments{
\item{xml}{object}
}
\value{
A vectors of characters with the names of document properties
}
\description{
Get the names of the document properties
}
