% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlFunctions.R
\name{get_types}
\alias{get_types}
\title{Get the xml types of the document properties}
\usage{
get_types(xml)
}
\arguments{
\item{xml}{object}
}
\value{
vector defined by factor(x, levels = c("i4", "lpwstr","filetime", "bool"), labels = c("numeric","character","date","logical") 
for types number, string, date and boolean.
}
\description{
Get the xml types of the document properties
}
