% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProperties.R
\name{loadProperties}
\alias{loadProperties}
\title{loadProperties}
\usage{
loadProperties(file)
}
\arguments{
\item{file}{A .docx file to read the data from}
}
\value{
A list of nodes
}
\description{
loadProperties
}
