% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProperties.R
\name{read_from_rdocx}
\alias{read_from_rdocx}
\title{Read custom.xml from rdocx object}
\usage{
read_from_rdocx(rdocx)
}
\arguments{
\item{rdocx}{object}
}
\value{
xml object
}
\description{
Read custom.xml from rdocx object
}
