% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveProperties.R
\name{save_to_rdocx}
\alias{save_to_rdocx}
\title{saveProperties}
\usage{
save_to_rdocx(rdocx, xml)
}
\arguments{
\item{file}{A .docx file to save the data to}

\item{properties}{The list to save}

\item{sourcefile}{A file to copy from and overwrite with the properties (Optional)}
}
\value{
A .docx file with the custom properties.
}
\description{
saveProperties
}
