% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveProperties.R
\name{saveProperties}
\alias{saveProperties}
\title{saveProperties}
\usage{
saveProperties(file, properties)
}
\arguments{
\item{file}{A .docx file to save the data to}

\item{properties}{The list to save}
}
\value{
The .docx file is updated with the data.
}
\description{
saveProperties
}
