% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlFunctions.R
\name{set_value}
\alias{set_value}
\title{set the value of a document property}
\usage{
set_value(xml, name, value)
}
\arguments{
\item{xml}{xml object}

\item{name}{name of property to change}

\item{value}{new value}
}
\value{
Nothing
}
\description{
set the value of a document property
}
