% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlFunctions.R
\name{xml_to_environment}
\alias{xml_to_environment}
\title{Convert Word custom xml object to an environment}
\usage{
xml_to_environment(xml)
}
\arguments{
\item{xml}{object to convert to environment}
}
\value{
a new environment
}
\description{
Convert Word custom xml object to an environment
}
