% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadProperties.R
\name{read_from_rdocx}
\alias{read_from_rdocx}
\title{loadProperties}
\usage{
read_from_rdocx(rdocx)
}
\arguments{
\item{file}{A .docx file to read the data from}
}
\value{
A list of nodes
}
\description{
loadProperties
}
